Create marketplace.sh
i have the marketplace publicly on here so that i can automatically update the script when i need to (e.g. adding more games to the list)
so that you wont have to manually update the entire program for it to work. also used to prevent script modifications on the client unless you change the script url